refactor(svelte2tsx): コンポーネントとスロットのハンドラを nodes/ へ切り出す
inline_component (Component / <svelte:component> / <svelte:self>)、
component_slots (名前付きスロットの子と $$slot_def 展開)、
slot_element (<slot> とスロット名・props ヘルパ) を純粋移動。
mod.rs は公開 API と型のみになり、中身が空になったバナーコメントを削除、
data-*/--* 属性ラップのテストを attributes/attribute.rs へ移設した。
出力・ロジックの変更なし。
