+++ crates/rsvelte_core/src/svelte2tsx/template/nodes/mod.rs
@@ -3,6 +3,7 @@
 pub(super) mod attach_tag;
 pub(super) mod await_block;
 pub(super) mod comment;
+pub(super) mod component_slots;
 pub(super) mod const_tag;
 pub(super) mod debug_tag;
 pub(super) mod declaration_tag;
@@ -10,10 +11,12 @@ pub(super) mod dynamic_element;
 pub(super) mod each_block;
 pub(super) mod element;
 pub(super) mod if_else_block;
+pub(super) mod inline_component;
 pub(super) mod key_block;
 pub(super) mod mustache_tag;
 pub(super) mod raw_mustache_tag;
 pub(super) mod render_tag;
+pub(super) mod slot_element;
 pub(super) mod snippet_block;
 pub(super) mod special_element;
 pub(super) mod text;
